commit visual studio project files
The current setup asumes protoc.exe is in the build path.
it also asumes the QTDIR environment variable is set.

I added 3 custom environment variables to control the location of libxml and zlib
EXTRALIBS_HEADERS -> dir to headers
EXTRALIBS_RELEASELIBS -> dir to libs to link in release mode
EXTRALIBS_DEBUGLIBS -> dir to libs to link in debug mode