GenWayAreaDat: Fix bug of losing some way areas
The way areas are written to file when either the rawWays or the nodeIds
fill the block size, or the last raw way is reached. However, the last
raw way might not be an way area (i.e. IsArea() returns false, or is in
blacklist). If so, all the way areas added after the last write will be
lost. Therefore, flushing the remaining way areas should be done outside
the loop.