D2DPainter: Remove dubious style-to-name maps from the DirectX renderer.
The whole point of the maps is far from being obvious, as the string
values are never used. Plus, one of the maps was never cleaned,
resulting in a memory leak.

All in all, the code looks more readable, is probably a bit faster since
it doesn't do any map lookup anymore, and is less leaky.