D2DPainter: remove the geometry caching
The cache wasn't working as intended, probably because of the hash
calculation. I'm removing it entirely instead of fixing it because I'm
not comfortable enough with the different data structures and the
rendering pipeline to understand when the cache is supposed to be
invalidated, and which objects actually have the same geometry.
Furthermore, tests on my end showed the cache removal didn't have a huge
impact on performance.