Issue #322: Removed POI flag for some types.
Definition of POI: Something that is located by name. POI is well known
but nobody wold address and doe sknow its actual address. The POI is
known because of singular instance and not because of its type. A POI
type has a rather small number of instances.

Thus a museum is a POI because it is address by its special name. An atm
or a bustop is definitely not a POI, becaue in general it would by selected
 by type and area but not because of its name. A hotel could be a POI
 because it is address by name but in some cases t is also address by
 type.