make taphandler slots virtual
We subclass the mapwidget to implement some special use cases. For example do we need to change the double tap behaviour to either zoom or to start some geocodeing at the clicked location depending on configuration by the user. Thus we need the slots to be virtual to be able to intercept.