Add a guard against null styles for areas
Apparently sometimes some areas don't have fill styles. I know at least
in one instance it's a bug since I can see a fill color in the default
style (which I used), but until I figure out why exactly this happens,
and because the other renderers all seem to do the same, I'm patching the
symptom before fixing the underlying cause.

For the record, the problem is when drawing an area in Rennes,France which has
the type "tourism_attraction_building".