new distance class, use it everywhere where distance is meter or kilometer
There is many places that use geographical distance or size in meters or kilometers.
Belong this, various methods is using distance in degrees or pixels.
To avoid errors when these units are mixed, new Distance class was created
for geographical distance. It can be easily converted to count of meters or kilometers,
comparison and math operators are provided to easy manipulation.
Internal presentation are always meters.