improve label layouting with multiple databases
It is done by delegating all label and icon layouting
to the last painter in the batch. With this simple trick
labels from multiple databases can't overlap. It is necessary
to use icon label name as a icon cache key, because icon ids
may be different for multiple StyleConfig instances.

Current implementation is just for the Qt map painter.