reuse FileScanner in LocationIndex using object pool
FileScanner opening may be expensive operation,
but LocationIndex may be recursive, because some Visit* method
may be called even from visitor. So it is not possible
use "global" scanner, because its position would be different
when returning from visitor. So, scanners are managed in
the ObjectPool that allows reusing the objects, and guarantee
exclusive access.

To keep small LocationIndex memory footprint, user may call
FlushCache method when index is not needed anymore.