fix VTable race condition in Worker
ProcessingLoop method is pure virtual in ThreadedWorker.
It may be called only after Vtable of child object is constructed.
For that reason it is responsibility for child class to call
Start method from its constructor. It is also beneficial
to mark child classes as final.

This commit is fixing call of pure virtual method.