fix search result comparison
`positionIt` is iterator to container with pointers.
To use it as reference for `Compare` lambda, it have to dereferenced twice.
Previous code was working, because `LocationEntry` constructor with
argument `(QObject* parent = nullptr)` was not explicit,
so this `LocationEntry` constructor was called implicitly with pointer
to original entry - it becomes a owner of a new entry.

`Compare` function is called with defaulted second entry,
comparison was not working obviously.

Moreover, entry is copied and transfered to QML with non-null parent,
so QML doesn't takes ownership of the entry. It may happen that
parent entry was deleted later in c++ code. In such case Qt deletes even
its children, that may be referenced from QML still.
Such scenario leads to crash.