fix UB in CoordBuffer::GenerateParallelWay
When two points in the way are on the same possition,
CoordBuffer::GenerateParallelWay leads to zero-division,
that is undefined. To fix, Normalize function should
do nothing when vector size is zero.

Result may not be correct (parallel way), when two identical points
are on the way start, but it is deterministic at least.