utf8helper: fix cached rawsize on transform
The rawsize member should be updated with the size in bytes of the stored code point.
Applying a transform operation could change the original code point to one with a different byte size. Therefore, the parser should store the size of the final code point resulting from the transform operation, not the size of the input code point.