Qt icon lookup module
IconLookup load map data asynchronously and lookup for objects that
contains symbol or icon and test if specific screen coordinate
contains icon / symbol. Without actual rendering of the map data.

When such icon/symbols is found, it is rendered, and object details
are exposed via asynchronous signal.

Note that it may detect different object than is visible on the map,
because label layouter algorithm works a bit differently.