respect rule of five in Projections
See https://en.cppreference.com/w/cpp/language/rule_of_three

When a base class is intended for polymorphic use,
its destructor may have to be declared public and virtual.
This blocks implicit moves (and deprecates implicit copies),
and so the special member functions have to be declared as defaulted.