lookup proper area's ring for search result details
Search indexes contains just object offset. But in case of the area,
matching may be just some ring, not whole area. Area may represent
lake with its own name for example, but on this lake may be islet
(inner Area ring) with different name...
So, we will try to find corresponding ring to get correct name,
type and alternative name.