use proper node when routable object is start or target
Current implementation of Qt Router module takes LocationEntry
center and lookup nearest routable node for route start or target.
It works fine in most cases, when start or target is POI or address
node. But When it is routable way (street) that is long and curved,
its bounding box center may lead to selecting node on different
routable object.

For example, when navigation is set to "Diskařská, Praha",
router is setup destination to different street:
https://www.openstreetmap.org/way/530888942#map=17/50.08187/14.38507

This change first tests if LocationEntry contains reference
to routable object, and select proper routable node. If it is not
routable, it continue with previous logic.