fix infinite loop on roundabout node postprocessing
Oneway entering roundabout is not considered
as a roundabout exit. But it is necessary to include
such oneways to postprocessing to be able find
the entering way to roundabout. We will get infinite
loop othervise, because we will never "enter" roundabout
during posprocessing node roundabouts (mini roundabout
in OSM terminology).