just print warning when multipolyton ring has wrong type
When multipolygon contains more than one level of nested
rings, second nested rigns should be outer. But is not
so big problem when even these rings are with inner type
(two nested inner rings). So, instead of skipping such
rings with invalid type, just log warning and continue.

When ring is skipped, ring grouping fail and whole
multipolygon is omitted from import. It is too strict.