Added the WellScoutedRoute application
Application was developed during https://www.fossgis.de/wiki/FOSSGIS_OSM_Communitytreffen_2023_Nummer_20

It tries to find a "good" roundtrip route (for biking or similar) using https://en.wikipedia.org/wiki/Bellman%E2%80%93Ford_algorithm based on the routing graph in the database.

Though currently fast, it is not fast enough to evaluate all options. The idea is,to have certain criteria and assumptions to further optimize the calculation.

Depending on current value of hard coded options it already generates good routes (though it is not clear for my why ad why not in other cases)

This is currently work in progress