better area optimization with "simple" constraint
When area/way is optimized (reduced geometry complexity),
there may be requirement that resulted geometry is simple
(no intersecting lines). This constraint is ensured by
searching for intersection and when there is some,
portion of the simplified geometry is removed.
In case of areas, such portion may be in the middle
or its begin and end. Geometric size should take into
account, not number of notes.